update tests move test to multiple files to easier to manage
